chore: refactor gamma folding for concat/slice RLC
